fix errors
Due to a change in the weights of the utility function, a different (but
totally fine) best substitution for test/model_p4.yaml was returned. The worker
next test failed for the same reason, another 'best' relation has been chosen
so the next() result was different.

Design error: The utility function should not influence the tests. It is
user-defined and the optimal substitution varies between models. The utility
function should have been set per test module/case, at least.