make 'provided' property useful
The 'provided' property was redundant, typically only provisions and constants
(i.e., the itoms) are specified in the model. Now, the property 'provided' can
be used to overwrite the value from the model (if variable in provision or
constants, then it is provided). This is used in the monitor which may not have
access to all itoms, so the provided variables can be specified in the model
via the node property 'provided'.